refactor(typography): move Typography component to UI directory and implement variant system
